rename files to avoid name conflicts between classes and modules
